enhanced integration tests
