+++ src/test/resources/reference.conf
@@ -10,3 +10,5 @@ akka {
 }
 
 mailo.cachingTTLSeconds = 60
+mailo.smtp.server = "0.0.0.0"
+mailo.smtp.port = "1025"
